refactor(flattenStylesheet): collapse CSS without Promise queues.
Remove the manual Promise queues, "compilationQueue" and "elementQueue", and
implemented them "automatically" with Observables. That is, RxJS and the
computer can keep Observables in order, even when their values may be
asynchronous.

By using "reply", we can turn a hot observable into a cold observable, with the
effect of starting these asynchronous actions, even before there's an attached
subscriber to the returned observable.